refactor(filter.go): move GetDirection, GetJSONBody
