+++ privacy-profile-composer/pkg/envoyfilter/internal/common/packet_processing.go
@@ -1,8 +1,13 @@
 package common
 
 import (
-	"github.com/envoyproxy/envoy/contrib/golang/common/go/api"
+	"encoding/json"
+	"fmt"
 	"log"
+	"net/url"
+	"strconv"
+
+	"github.com/envoyproxy/envoy/contrib/golang/common/go/api"
 )
 
 const anyPurpose = "ANY"
@@ -20,6 +25,13 @@ type HeaderMetadata struct {
 	SvcName string
 	Purpose string
 }
+
+type SidecarDirection int
+
+const (
+	Inbound SidecarDirection = iota
+	Outbound
+)
 
 func LogDecodeHeaderData(header api.RequestHeaderMap) {
 	log.Printf("%v (%v) %v://%v%v\n", header.Method(), header.Protocol(), header.Scheme(), header.Host(), header.Path())
@@ -84,4 +96,56 @@ func ExtractHeaderData(header api.RequestHeaderMap) HeaderMetadata {
 	}
 
 	return metadata
+}
+
+func GetDirection(callbacks api.FilterCallbackHandler) (SidecarDirection, error) {
+	directionEnum, err := callbacks.GetProperty("xds.listener_direction")
+	if err != nil {
+		return -1, fmt.Errorf("cannot determine sidecar direction as there is no xds.listener_direction key")
+	}
+	directionInt, err := strconv.Atoi(directionEnum)
+	if err != nil {
+		// check https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/core/v3/base.proto#envoy-v3-api-enum-config-core-v3-trafficdirection
+		return -1, fmt.Errorf("envoy's xds.listener_direction key does not contain an integer " +
+			"check the Envoy docs for the range of values for this key")
+	}
+
+	if directionInt == 0 {
+		return -1, fmt.Errorf("envoy's xds.listener_direction key indicates that this sidecar is deployed as a gateway." +
+			"Prose does not need to be run in a gateway sidecar." +
+			"It will continue to get deployed in other sidecars that are configured as inbound or outbound sidecars")
+	}
+
+	if directionInt == 1 {
+		return Inbound, nil
+	}
+	if directionInt == 2 {
+		return Outbound, nil
+	}
+
+	return -1, fmt.Errorf("envoy's xds.listener_direction key contains an unsupported value for the direction enum: %d "+
+		"check the Envoy docs for the range of values for this key", directionInt)
+}
+
+func GetJSONBody(headerMetadata HeaderMetadata, buffer api.BufferInstance) ([]byte, error) {
+	var jsonBody []byte
+
+	if headerMetadata.ContentType == nil {
+		return nil, fmt.Errorf("ContentType header is not set. Cannot analyze body")
+	} else if *headerMetadata.ContentType == "application/x-www-form-urlencoded" {
+		query, err := url.ParseQuery(buffer.String())
+		if err != nil {
+			return nil, fmt.Errorf("Failed to start decoding JSON data")
+		}
+		log.Println("  <<decoded x-www-form-urlencoded data: ", query)
+		jsonBody, err = json.Marshal(query)
+		if err != nil {
+			return nil, fmt.Errorf("Could not transform URL encoded data to JSON to pass to Presidio")
+		}
+	} else if *headerMetadata.ContentType == "application/json" {
+		jsonBody = buffer.Bytes()
+	} else {
+		return nil, fmt.Errorf("Cannot analyze a body with contentType '%s'\n", *headerMetadata.ContentType)
+	}
+	return jsonBody, nil
 }
